Dashboard refactor
