+++ pages/user/users.js
@@ -1,10 +1,17 @@
-import { useEffect, useState } from "react"
-import { API_BASE_URL } from '../index';
+import React, { useEffect, useState } from "react"
+import { DataGrid, GridToolbar  } from '@mui/x-data-grid'
 
+const API_BASE_URL = process.env.NEXT_PUBLIC_API_URL;
 
+const columns = [
+  { field: 'username', headerName: 'User Name', width: 130 },
+  { field: 'url', headerName: 'URL', width: 130 },
+]
 
 export default function Home() {
     const [data, setData] = useState([])
+    const [filteredData,setFilteredData] = React.useState("")
+
     useEffect(()=>{
         fetch(`${API_BASE_URL}/users`).then((res)=>{
             return res.json()
@@ -14,12 +21,26 @@ export default function Home() {
         })
     },[])
     return(
-        <div>
-            {
-                data.map((user)=> (
-                    <li key={`user.name`}>{user.name}</li>
-                ))
-            }
-        </div>
+        <React.Fragment>
+            <DataGrid 
+                {...data}  
+                getRowId={(row) => row.username}
+                rows={data}
+                columns={columns}
+                initialState={{
+                pagination: {
+                    paginationModel: { page: 0, pageSize: 5 },
+                },
+                }}
+                pageSizeOptions={[5, 10]}
+                checkboxSelection
+                slots={{ toolbar: GridToolbar }}
+                slotProps={{
+                  toolbar: {
+                    showQuickFilter: true,
+                  },
+                }}
+            />
+        </React.Fragment>
     )
 }
